requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchaudio==2.1.0
openai-whisper==20231117
pydantic==2.5.0
python-multipart==0.0.6
